members.serializer Refactoring : Separate modules
