+++ app/members/serializers/__init__.py
@@ -1,5 +1,11 @@
-from .auth import *
 from .facebook import *
 from .facebook_auth import *
-from .image import *
+from .user_profile import *
 from .user import *
+from .user_create import *
+from .user_update import *
+
+
+# 여러번 테스트 결과 __init__.py에서 import 하는 모듈이 import하려는
+# 모듈보다 위에있어서 에러가 난 것으로 추측.
+# 정확한 이유는 불명
