--- app/members/serializers/image.py
@@ -1,15 +1,0 @@
-# from rest_framework import serializers
-#
-# from members.models import UserProfileImages
-#
-#
-# class UserProfileImagesSerializer(serializers.ModelSerializer):
-#
-#     # img_profile = serializers.ImageField(read_only=True)
-#
-#     class Meta:
-#         model = UserProfileImages
-#         # fields = (
-#         #     'img_profile',
-#         # )
-#         fields = '__all__'
