+++ app/members/serializers/user.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model
 from rest_framework import serializers
 
-from members.serializers import UserProfileImagesSerializer
+from ..serializers import UserProfileImagesSerializer
 
 User = get_user_model()
